feat(PaymentFactory): static factory methods
- Payment 생성 책임이 커짐에 따라 PaymentFactory 객체로 책임 분리
